+++ build.cmd
@@ -1,1 +1,2 @@
 pyinstaller --onefile --hidden-import libs.OrderWork.OrderWorker  mainapp.py
+pyinstaller --onefile mainapp.py REM just use this
