+++ requirements.txt
@@ -3,19 +3,26 @@ aiosignal==1.3.1
 altgraph==0.17.4
 async-timeout==4.0.3
 attrs==23.2.0
+certifi==2024.2.2
+charset-normalizer==3.3.2
 frozenlist==1.4.1
 idna==3.6
+importlib_metadata==7.1.0
+macholib==1.16.3
 multidict==6.0.5
 packaging==24.0
 pefile==2023.2.7
+pybit==5.6.2
+pycryptodome==3.19.0
 pyinstaller==5.13.0
 pyinstaller-hooks-contrib==2024.3
 PyQt5==5.15.10
 PyQt5-Qt5==5.15.2
 PyQt5-sip==12.13.0
 pywin32-ctypes==0.2.2
+requests==2.31.0
+urllib3==2.2.1
+websocket-client==1.5.0
 websockets==12.0
 yarl==1.9.4
-requests>=2.22.0
-websocket-client==1.5.0
-pycryptodome==3.19.0
+zipp==3.18.1
